Promote globe view from the sandbox to the home page
