+++ next.config.mjs
@@ -1,4 +1,13 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  async rewrites() {
+    return [
+      {
+        source: "/area/:areaId",
+        destination: "/",
+      },
+    ];
+  },
+};
 
 export default nextConfig;
