+++ src/app/globe/components/country-card.tsx
