+++ src/app/globe/components/globe-panel.tsx
