+++ src/app/globe/state/machine.ts
