+++ src/app/page.tsx
@@ -1,23 +1,7 @@
-import React from "react";
-import LinkButton from "./components/link-button";
+import Globe from "./globe";
 
-const currentPages = [
-  { href: "/about", label: "About" },
-  { href: "/sandbox/simple-map", label: "Simple Map" },
-  { href: "/sandbox/globe-view", label: "Globe View" },
-  { href: "/sandbox/mvt-test", label: "MVT View" },
-];
+const Home = () => {
+  return <Globe />;
+};
 
-export default function Home() {
-  return (
-    <div className="flex h-screen items-center justify-center">
-      <div className="flex space-x-4">
-        {currentPages.map((page) => (
-          <LinkButton key={page.href} href={page.href}>
-            {page.label}
-          </LinkButton>
-        ))}
-      </div>
-    </div>
-  );
-}
+export default Home;
